Suggest correct file extension when cropping pdf, djvu and tiff
Previously this was suggesting a .jpg extension for multipage
files and keeping the same extension for everything else.

However, single page .pdf and .djvu should still have the .jpg
extension. Multipage .tiff files should keep .tiff and not be .jpg

This was causing upload failures unless the user noticed and
changes the extension.